Move buffers package to internal, add some test coverage
