+++ zapcore/entry.go
@@ -26,7 +26,7 @@ import (
 	"sync"
 	"time"
 
-	"go.uber.org/zap/buffers"
+	"go.uber.org/zap/internal/buffers"
 	"go.uber.org/zap/internal/exit"
 )
